edit LemmaService.

+++ src/main/java/searchengine/services/SearchService.java
@@ -3,5 +3,5 @@ package searchengine.services;
 import org.springframework.http.ResponseEntity;
 
 public interface SearchService {
-    ResponseEntity search(String query, String siteUrl, long limit, long offset);
+    ResponseEntity search(String query, String siteUrl, int limit, int offset);
 }
